fix: redir refactor in progress
