+++ includes/mini_shell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   mini_shell.h                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tklimova <tklimova@student.42berlin.de>    +#+  +:+       +#+        */
+/*   By: tklimova <tklimova@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/05 14:58:51 by tklimova          #+#    #+#             */
-/*   Updated: 2023/10/22 21:33:03 by tklimova         ###   ########.fr       */
+/*   Updated: 2023/10/24 15:25:32 by tklimova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -155,6 +155,11 @@ void			syntax_parser(t_data *data);
 void			init_exec_data(t_exec_data *exec_data);
 
 void			reset_std(t_exec_data *exec_data, int fd);
+
+void			here_doc(t_exec_data *exec_data, t_redir_data *redir_data,
+					int *prev_fd);
+
+void 			clear_exec_data(t_exec_data *exec_data, t_data *data);
 
 void			make_redirections(t_parser_data *parser_node,
 					t_exec_data *exec_data, int *prev_fd);
