+++ src/executor/redirect.c
@@ -3,57 +3,26 @@
 /*                                                        :::      ::::::::   */
 /*   redirect.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tklimova <tklimova@student.42berlin.de>    +#+  +:+       +#+        */
+/*   By: tklimova <tklimova@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/18 15:59:55 by tklimova          #+#    #+#             */
-/*   Updated: 2023/10/22 21:33:39 by tklimova         ###   ########.fr       */
+/*   Updated: 2023/10/24 15:25:38 by tklimova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/mini_shell.h"
 
-void	here_doc(t_exec_data *exec_data, t_redir_data *redir_data, int *prev_fd)
-{
-	char	*buffer;
-	int		len[2];
-	int		fd[2];
-
-	if (!redir_data->is_here_doc)
-		return;
-	len[0] = ft_strlen(redir_data->text);
-	if (pipe(fd) == -1)
-		return;
-	buffer = malloc(sizeof(char) * (len[0] + 2));
-	if (!buffer)
-		return;
-	len[1] = read(exec_data->stdin_dup, buffer, len[0]);
-	buffer[len[1]] = '\0';
-	while (ft_strcmp(redir_data->text, buffer) && len[1] > 0)
-	{
-		write(fd[1], buffer, len[1]);
-		len[1] = read(exec_data->stdin_dup, buffer, len[0]);
-		buffer[len[1]] = '\0';
-	}
-	close(fd[1]);
-	if (!exec_data->status_code)
-		*prev_fd = fd[0];
-		// dup2(fd[0], STDIN_FILENO);
-	else
-		close(fd[0]);
-	free(buffer);
-}
-
 void	open_write_stream(t_redir_data *redir_data,
 		t_exec_data *exec_data)
 {
-		if (exec_data->status_code)
-			return;
-		if (exec_data->was_stdoutredir)
-			close(exec_data->fd_out);
-		exec_data->was_stdoutredir = 1;
-		exec_data->fd_out = open(redir_data->text, redir_data->flags, 0755);
-		if (exec_data->fd_out < 0)
-			exec_data->status_code = 1;
+	if (exec_data->status_code)
+		return;
+	if (exec_data->was_stdoutredir)
+		close(exec_data->fd_out);
+	exec_data->was_stdoutredir = 1;
+	exec_data->fd_out = open(redir_data->text, redir_data->flags, 0755);
+	if (exec_data->fd_out < 0)
+		exec_data->status_code = 1;
 }
 
 void	open_stream(t_redir_data *redir_data,
@@ -72,19 +41,15 @@ void	open_stream(t_redir_data *redir_data,
 		close(*prev_fd);
 		exec_data->was_stdinredir = 1;
 		*prev_fd = open(redir_data->text, redir_data->flags);
+		if (*prev_fd < 0)
+		{
+			if (exec_data->was_stdoutredir)
+				close(exec_data->fd_out);
+			exec_data->status_code = 1;
+		}
 	}
 	else
-	{
 		open_write_stream(redir_data, exec_data);
-	}
-	if (!redir_data->std_fd && *prev_fd < 0)
-	{
-		if (exec_data->was_stdoutredir)
-			close(exec_data->fd_out);
-		exec_data->status_code = 1;
-	}
-	// else if (!redir_data->std_fd)
-	// 	close(fd[redir_data->std_fd]);
 }
 
 void update_exec_data(t_parser_data *parser_node, t_exec_data *exec_data)
